# src/controllers/admin/admin.controller.js
import { BaseController } from '#controllers/base.controller.js';
import { HTTP_STATUS } from '#constants';
import { adminValidator, auditAdminAction, validate } from '#middlewares';
import {
  getAllChallengesScheme,
  updateChallengeStatusScheme,
} from './dto/admin.dto.js';
export class AdminController extends BaseController {
  #challengesService;
  constructor({ challengesService }) {
    super();
    this.#challengesService = challengesService;
  }

  routes() {
    this.router.get(
      '/challenges',
      adminValidator,
      auditAdminAction,
      validate('query', getAllChallengesScheme),
      (req, res, next) => this.getAllChallenges(req, res, next),
    );
    this.router.get(
      '/challenges/:challengeId',
      adminValidator,
      auditAdminAction,
      (req, res, next) => this.getChallengeDetailById(req, res, next),
    );
    this.router.patch(
      '/challenges/:challengeId',
      adminValidator,
      auditAdminAction,
      validate('params', updateChallengeStatusScheme.params),
      validate('body', updateChallengeStatusScheme.body),
      (req, res, next) => this.updateChallengeStatus(req, res, next),
    );

    return this.router;
  }

  async getAllChallenges(req, res) {
    const { page, pageSize, sort, keyword } = req.query;
    const result = await this.#challengesService.getAllChallenges({
      page: Number(page) || 1,
      pageSize: Number(pageSize) || 10,
      sort,
      keyword,
      actor: req.user,
    });

    res.status(HTTP_STATUS.OK).json(result);
  }

  async getChallengeDetailById(req, res) {
    const { challengeId } = req.params;
    const result = await this.#challengesService.getChallengeDetailById(
      challengeId,
      req.user,
    );

    if (!result) {
      return res
        .status(HTTP_STATUS.NOT_FOUND)
        .json({ message: '해당 챌린지를 찾을 수 없습니다.' });
    }

    res.status(HTTP_STATUS.OK).json(result);
  }

  async updateChallengeStatus(req, res) {
    const challengeId = req.params.challengeId;
    const data = req.body;

    const result = await this.#challengesService.updateChallengeStatus(
      challengeId,
      data,
      req.user,
    );

    res.status(HTTP_STATUS.OK).json({ result });
  }
}