# swagger.yaml
openapi: 3.0.3
info:
  title: DocThru API
  version: 1.0.0
  description: |
    DocThru 백엔드 API 명세 초안
    - 현재 확정된 Prisma 스키마 기준
servers:
  - url: http://localhost:5001/api
...
paths:
  /admin/management:
    get:
      tags: [Admin]
      summary: 챌린지 관리 목록 조회 (승인 대기 챌린지 등)
      security:
        - accessTokenCookie: []
      parameters:
        - $ref: '#/components/parameters/CursorQuery'
        - $ref: '#/components/parameters/LimitQuery'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/ChallengeStatus'
      responses:
        '200':
          description: 관리 목록 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /auth/signup:
    post:
      tags: [Auth]
      summary: 회원가입
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: 회원가입 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags: [Auth]
      summary: 로그인
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 로그인 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/password-reset/request:
    post:
      tags: [Auth]
      summary: 비밀번호 재설정 메일 요청
      description: 계정 존재 여부와 무관하게 동일한 응답을 반환합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequestBody'
      responses:
        '200':
          description: 요청 접수(항상 동일 메시지)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordResetRequestResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          description: 요청 한도 초과
  /auth/password-reset/confirm:
    post:
      tags: [Auth]
      summary: 비밀번호 재설정 확정(토큰 1회용)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirmBody'
      responses:
        '200':
          description: 비밀번호 변경 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageOnlyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          description: 요청 한도 초과
  /auth/logout:
    post:
      tags: [Auth]
      summary: 현재 기기 로그아웃
      responses:
        '200':
          description: 로그아웃 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageOnlyResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout/all:
    post:
      tags: [Auth]
      summary: 모든 기기 로그아웃
      responses:
        '200':
          description: 모든 기기 로그아웃 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageOnlyResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: access token 재발급
      responses:
        '200':
          description: 토큰 재발급 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/{provider}/login:
    get:
      tags: [Auth]
      summary: OAuth 로그인 시작
      parameters:
        - $ref: '#/components/parameters/ProviderParam'
      responses:
        '302':
          description: OAuth 제공자 로그인 페이지로 리다이렉트
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/{provider}/callback:
    get:
      tags: [Auth]
      summary: OAuth 인증 처리
      parameters:
        - $ref: '#/components/parameters/ProviderParam'
        - in: query
          name: code
          required: true
          schema:
            type: string
        - in: query
          name: state
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OAuth 인증 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '302':
          description: 프론트 페이지로 리다이렉트
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /challenges:
    get:
      tags: [Challenges]
      summary: 전체 챌린지 목록 조회
      parameters:
        - $ref: '#/components/parameters/CursorQuery'
        - $ref: '#/components/parameters/LimitQuery'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/ChallengeStatus'
        - in: query
          name: category
          schema:
            $ref: '#/components/schemas/Category'
        - in: query
          name: type
          schema:
            $ref: '#/components/schemas/Type'
        - in: query
          name: keyword
          schema:
            type: string
      responses:
        '200':
          description: 챌린지 목록 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

    post:
      tags: [Challenges]
      summary: 새 챌린지 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChallengeRequest'
      responses:
        '201':
          description: 챌린지 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /challenges/me:
    get:
      tags: [Challenges]
      summary: 내가 신청한 챌린지 목록 조회
      parameters:
        - $ref: '#/components/parameters/CursorQuery'
        - $ref: '#/components/parameters/LimitQuery'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/ApplicationStatus'
      responses:
        '200':
          description: 내가 신청한 챌린지 목록 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /challenges/{challengeId}:
    get:
      tags: [Challenges]
      summary: 특정 챌린지 상세 조회
      parameters:
        - $ref: '#/components/parameters/ChallengeIdParam'
      responses:
        '200':
          description: 챌린지 상세 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Challenges]
      summary: 특정 챌린지 정보 수정
      parameters:
        - $ref: '#/components/parameters/ChallengeIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateChallengeRequest'
      responses:
        '200':
          description: 챌린지 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

    delete:
      tags: [Challenges]
      summary: 특정 챌린지 삭제
      parameters:
        - $ref: '#/components/parameters/ChallengeIdParam'
      responses:
        '200':
          description: 챌린지 삭제 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageOnlyResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /works:
    get:
      tags: [Works]
      summary: 작업물 목록 조회
      parameters:
        - in: query
          name: challengeId
          required: false
          schema:
            type: string
            description: 특정 챌린지의 작업물만 조회할 때 사용
        - $ref: '#/components/parameters/CursorQuery'
        - $ref: '#/components/parameters/LimitQuery'
        - in: query
          name: sort
          schema:
            type: string
            example: latest
      responses:
        '200':
          description: 작업물 목록 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkListResponse'

    post:
      tags: [Works]
      summary: 작업물 등록
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkRequest'
      responses:
        '201':
          description: 작업물 등록 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /works/{workId}:
    get:
      tags: [Works]
      summary: 작업물 상세 조회
      parameters:
        - $ref: '#/components/parameters/WorkIdParam'
      responses:
        '200':
          description: 작업물 상세 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Works]
      summary: 작업물 수정
      parameters:
        - $ref: '#/components/parameters/WorkIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkRequest'
      responses:
        '200':
          description: 작업물 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Works]
      summary: 작업물 삭제
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WorkIdParam'
      responses:
        '200':
          description: 작업물 삭제 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageOnlyResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    accessTokenCookie:
      type: apiKey
      in: cookie
      name: accessToken
      description: 로그인 시 발급되는 Access Token 쿠키
  parameters:
    ProviderParam:
      in: path
      name: provider
      required: true
      schema:
        $ref: '#/components/schemas/AuthProvider'
...
    ChallengeIdParam:
      in: path
      name: challengeId
      required: true
      schema:
        type: string
        example: 01HV123ABC456DEF789GHIJKL
    WorkIdParam:
      in: path
      name: workId
      required: true
      schema:
        type: string
        example: 01HV123ABC456DEF789GHIJKL
...
    CursorQuery:
      in: query
      name: cursor
      required: false
      schema:
        type: string
        description: 다음 페이지 조회를 위한 커서(ULID)
    LimitQuery:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: 잘못된 요청입니다.
            code: BAD_REQUEST
            details: null
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: 인증이 필요합니다.
            code: UNAUTHORIZED
            details: null
    Forbidden:
      description: 권한 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: 접근 권한이 없습니다.
            code: FORBIDDEN
            details: null
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: 리소스를 찾을 수 없습니다.
            code: RESOURCE_NOT_FOUND
            details: null
    Conflict:
      description: 충돌 발생
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: 이미 존재하는 리소스입니다.
            code: RESOURCE_CONFLICT
            details: null
  schemas:
    Role:
      type: string
      enum:
        - USER
        - ADMIN
    Grade:
      type: string
      enum:
        - NORMAL
        - EXPERT
    AuthProvider:
      type: string
      enum:
        - GOOGLE
...
    ChallengeStatus:
      type: string
      enum:
        - PENDING
        - APPROVED
        - REJECTED
        - DELETED
    Category:
      type: string
      enum:
        - DOCUMENT
        - BLOG
    Type:
      type: string
      enum:
        - NEXT_JS
        - API
        - CAREER
        - MODERN_JS
        - WEB
    ApplicationStatus:
      type: string
      enum:
        - PENDING
        - APPROVED
        - REJECTED
    MessageOnlyResponse:
      type: object
      properties:
        message:
          type: string
          example: 성공했습니다.
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: 에러 메시지
        code:
          type: string
          description: 에러 코드
        details:
          type: object
          nullable: true
          description: 필드별 에러 상세 정보 (검증 실패 등)
    Pagination:
      type: object
      properties:
        nextCursor:
          type: string
          nullable: true
        hasNext:
          type: boolean
    User:
      type: object
      properties:
        id:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        grade:
          $ref: '#/components/schemas/Grade'
        email:
          type: string
          format: email
        nickname:
          type: string
        image:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
...
    UserSummary:
      type: object
      properties:
        id:
          type: string
        nickname:
          type: string
        image:
          type: string
          nullable: true
...
    Challenge:
      type: object
      properties:
        id:
          type: string
        authorId:
          type: string
        title:
          type: string
        originalUrl:
          type: string
          format: uri
        type:
          $ref: '#/components/schemas/Type'
        category:
          $ref: '#/components/schemas/Category'
        description:
          type: string
        deadline:
          type: string
          format: date-time
        maxParticipants:
          type: integer
        status:
          $ref: '#/components/schemas/ChallengeStatus'
        isClosed:
          type: boolean
        declineReason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
        author:
          $ref: '#/components/schemas/UserSummary'
...
    Work:
      type: object
      properties:
        id:
          type: string
        challengeId:
          type: string
        participantId:
          type: string
          nullable: true
        userId:
          type: string
          nullable: true
        content:
          type: string
          description: TipTap JSON 문자열 또는 일반 텍스트
        likeCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/UserSummary'
...
    SignupRequest:
      type: object
      required:
        - email
        - password
        - nickname
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        nickname:
          type: string
          minLength: 2
          maxLength: 20
        image:
          type: string
          format: uri
          nullable: true
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    PasswordResetRequestBody:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
    PasswordResetRequestResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: 요청이 접수되었습니다. 등록된 이메일이 있으면 안내를 보냅니다.
    PasswordResetConfirmBody:
      type: object
      required:
        - token
        - password
      properties:
        token:
          type: string
          description: 이메일로 받은 일회용 토큰(평문은 서버에 저장하지 않음)
        password:
          type: string
          minLength: 8
...
    CreateChallengeRequest:
      type: object
      required:
        - title
        - originalUrl
        - type
        - category
        - description
        - deadline
        - maxParticipants
      properties:
        title:
          type: string
          maxLength: 100
        originalUrl:
          type: string
          format: uri
        type:
          $ref: '#/components/schemas/Type'
        category:
          $ref: '#/components/schemas/Category'
        description:
          type: string
        deadline:
          type: string
          format: date-time
        maxParticipants:
          type: integer
          minimum: 1
    UpdateChallengeRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 100
        originalUrl:
          type: string
          format: uri
        type:
          $ref: '#/components/schemas/Type'
        category:
          $ref: '#/components/schemas/Category'
        description:
          type: string
        deadline:
          type: string
          format: date-time
        maxParticipants:
          type: integer
          minimum: 1
        status:
          $ref: '#/components/schemas/ChallengeStatus'
        declineReason:
          type: string
          nullable: true
...
    CreateWorkRequest:
      type: object
      required:
        - challengeId
        - content
      properties:
        challengeId:
          type: string
        content:
          type: string
          description: TipTap JSON 문자열
    UpdateWorkRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          description: TipTap JSON 문자열
...
    LoginResponse:
      type: object
      properties:
        message:
          type: string
          example: 로그인 성공
        data:
          type: object
          properties:
            accessToken:
              type: string
            user:
              $ref: '#/components/schemas/User'
    RefreshTokenResponse:
      type: object
      properties:
        message:
          type: string
          example: 토큰 재발급 성공
        data:
          type: object
          properties:
            accessToken:
              type: string
    UserResponse:
      type: object
      properties:
        message:
          type: string
        data:
          $ref: '#/components/schemas/User'
...
    ChallengeDetailResponse:
      type: object
      properties:
        message:
          type: string
        data:
          $ref: '#/components/schemas/Challenge'
    ChallengeListResponse:
      type: object
      properties:
        message:
          type: string
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Challenge'
            pagination:
              $ref: '#/components/schemas/Pagination'
...
    WorkDetailResponse:
      type: object
      properties:
        message:
          type: string
        data:
          $ref: '#/components/schemas/Work'
    WorkListResponse:
      type: object
      properties:
        message:
          type: string
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Work'
            pagination:
              $ref: '#/components/schemas/Pagination'
...